Split ImportScriptModal into logic and presentation components
